+++ MechanoCraft/MechanoCraft.cs
@@ -4,6 +4,10 @@ using Microsoft.Xna.Framework.Input;
 using System.Collections.Generic;
 using MonoGame.Extended.Entities;
 using MonoGame.Extended;
+using MechanoCraft.Input;
+using MechanoCraft.Render;
+using MonoGame.Extended.Sprites;
+using MonoGame.Extended.TextureAtlases;
 
 namespace MechanoCraft
 {
@@ -12,8 +16,7 @@ namespace MechanoCraft
         private GraphicsDeviceManager _graphics;
         private SpriteBatch _spriteBatch;
         private World _world;
-        private Dictionary<SpriteRenderer, Vector2> spriteRenderers = new Dictionary<SpriteRenderer, Vector2>();
-        private MouseState oldState;
+
         public MechanoCraft()
         {
             _graphics = new GraphicsDeviceManager(this);     
@@ -25,21 +28,20 @@ namespace MechanoCraft
         {
 
             // TODO: Add your initialization logic here
-            _world = new WorldBuilder().AddSystem(new RenderSystem(GraphicsDevice)).Build();
+            _world = new WorldBuilder().AddSystem(new SpriteRenderSystem(GraphicsDevice)).Build();
             Components.Add(_world);
-            spriteRenderers.Add(new SpriteRenderer(Content.Load<Texture2D>("minerai"), Color.White), new Vector2(10, 10));
-            spriteRenderers.Add(new SpriteRenderer(Content.Load<Texture2D>("MinerShadow"), Color.White), new Vector2(10, 10));
-            spriteRenderers.Add(new SpriteRenderer(Content.Load<Texture2D>("Miner"), Color.White), new Vector2(10, 10));
-            spriteRenderers.Add(new SpriteRenderer(Content.Load<Texture2D>("crafter"), Color.White), new Vector2(80, 10));
-            spriteRenderers.Add(new SpriteRenderer(Content.Load<Texture2D>("polish"), Color.White), new Vector2(60, 10));
-            spriteRenderers.Add(new SpriteRenderer(Content.Load<Texture2D>("smelter"), Color.White), new Vector2(40, 10));
-            foreach (var spriteRenderer in spriteRenderers)
+            Entity entity = _world.CreateEntity();
+            InputHandler.GetInstance().AddInputListener(Keys.Space, () =>
             {
-                var entity = _world.CreateEntity();
-                entity.Attach(new Transform2(spriteRenderer.Value));
-                entity.Attach(spriteRenderer.Key);
-            }
-            spriteRenderers.Clear();
+                Entity entity1 = _world.CreateEntity();
+                entity1.Attach( new Transform2());
+                entity1.Attach( new Sprite(Content.Load<Texture2D>("Crafter")));
+                entity1.Get<Transform2>().Position = new Vector2(Mouse.GetState().X, Mouse.GetState().Y);
+            });
+            InputHandler.GetInstance().AddInputListener(Keys.Escape, () =>
+            {
+                Exit();
+            });
             base.Initialize();
         }
 
@@ -51,29 +53,9 @@ namespace MechanoCraft
 
         protected override void Update(GameTime gameTime)
         {
-            if (GamePad.GetState(PlayerIndex.One).Buttons.Back == ButtonState.Pressed || Keyboard.GetState().IsKeyDown(Keys.Escape))
-                Exit();
-
-
-            MouseState newState = Mouse.GetState();
-
-            if (newState.LeftButton == ButtonState.Pressed && oldState.LeftButton == ButtonState.Released)
-            {
-                spriteRenderers.Add(new SpriteRenderer(Content.Load<Texture2D>("MinerShadow"), Color.White), new Vector2(newState.Position.X, newState.Position.Y));
-                spriteRenderers.Add(new SpriteRenderer(Content.Load<Texture2D>("Miner"), Color.White), new Vector2(newState.Position.X, newState.Position.Y));
-                foreach (var spriteRenderer in spriteRenderers)
-                {
-                    spriteRenderer.Key.drawMode = SpriteRenderer.DrawMode.Center;
-                    var entity = _world.CreateEntity();
-                    entity.Attach(new Transform2(spriteRenderer.Value));
-                    entity.Attach(spriteRenderer.Key);
-                }
-                spriteRenderers.Clear();
-            }
-
-            oldState = newState;
             // TODO: Add your update logic here
             _world.Update(gameTime);
+            InputHandler.GetInstance().ProcessListeners();
             base.Update(gameTime);
         }
